exemplo-jwt/requirements.txt
django
djangorestframework
djangorestframework-jwt
psycopg2-binary
requests